Fix sending more than one followup input
The bug was that we were sending `content: [{ text: '...' }]`, where we also
needed to send `type` in that inner object if doing this. For now, I've
setup parsing so that it translates this to `content: '...'` which is a bit
simpler, with the idea to come back to this once I've hooked up other providers
to hopefully make something that's a bit more portable/cross-provider.

Signed-off-by: Matthew Peveler <mpeveler@tigerdata.com>